requirements.txt
mediapipe==0.10.14
numpy==2.1.1
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
Pillow==10.4.0
Requests==2.32.3
selenium==4.24.0
deepface==0.0.93
tf-keras==2.17.0